rkwifibt: ap6203bm: update fw/tb_start_wifi.sh
1. optimize getting dhcp time
2. optimize anti-interference performance
3. support getting ssid/psk/ip/dns info for vendor

+++ tb_start_wifi.sh
@@ -3,7 +3,11 @@
 SSID=
 WIFISSID=$1
 WIFIPWD=$2
+VSSID=
+VPWD=
+VVAILD=
 IPADDR=
+NETMASK=
 WPAPID=""
 NETMASK=
 GW=
@@ -16,8 +20,47 @@ NEW_SSID=
 getdhcp_cnt=0
 connect_cnt=0
 wpa_cnt=0
+NET_NEW_CONFIG=0
+IPADDR_TEST=
+tcpka_cnt=0
 
 trap "echo 'ifconfig wlan0 down'; ifconfig wlan0 down; exit" SIGTERM
+
+function udhcpcd() {
+	echo "udhcpc ..."
+
+	if [ "$NET_NEW_CONFIG" == "0" ];then
+		echo "get dhcp info from vendor"
+		IPADDR=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $4}'`
+		NETMASK=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $5}'`
+		GW=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $6}'`
+		DNS=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $7}'`
+
+		ifconfig wlan0 $IPADDR netmask $NETMASK
+		route add default gw $GW
+		echo "nameserver $DNS" > /etc/resolv.conf
+		echo $IPADDR $NETMASK $GW $DNS
+		wl PM 0
+		exit
+	fi
+
+	echo "get dhcp by udhcpc"
+	udhcpc -i wlan0 -t 15 -q
+
+	IPADDR_TEST=`ifconfig wlan0 | grep inet | awk '{print $2}' | awk -F: '{print $2}' | awk -F. '{print $1}'`
+	if [ "$IPADDR_TEST" != "" ] && [ "$IPADDR_TEST" != "127" ] && [ "$IPADDR_TEST" != "169" ];then
+		IPADDR=`ifconfig wlan0 | grep inet | awk '{print $2}' | awk -F: '{print $2}'`
+		NETMASK=`ifconfig wlan0 | grep inet | awk '{print $4}' | awk -F: '{print $2}'`
+		GW=`route -n | awk '{print $2}' | sed -n '3p'`
+		DNS=`cat /etc/resolv.conf | grep nameserver | awk '{print $2}'`
+		echo "ip_info: $IPADDR $NETMASK $GW $DNS"
+		vendor_storage -w VENDOR_CUSTOM_ID_1E -t string -i 1,$WIFISSID,$WIFIPWD,$IPADDR,$NETMASK,$GW,$DNS
+		wl PM 0
+	else
+		echo "dhcp failed !!!"
+		exit
+	fi
+}
 
 function getdhcp() {
 	while true
@@ -58,6 +101,7 @@ function getdhcp() {
 }
 
 function check_wlan0() {
+	echo "check wlan0 ..."
 	while true
 	do
 		lsmod | grep bcmdhd
@@ -70,7 +114,7 @@ function check_wlan0() {
 		fi
 
 		let insmod_cnt++
-		if [ $insmod_cnt -gt 10 ]; then
+		if [ $insmod_cnt -gt 3 ]; then
 			echo "insmod failed !!!"
 			echo 0 > /sys/class/leds/blue/brightness
 			exit
@@ -89,7 +133,7 @@ function check_wlan0() {
 		fi
 
 		let check_cnt++
-		if [ $check_cnt -gt 10 ]; then
+		if [ $check_cnt -gt 3]; then
 			echo "check wlan0 failed!!!"
 			echo 0 > /sys/class/leds/blue/brightness
 			exit
@@ -98,6 +142,7 @@ function check_wlan0() {
 }
 
 function tcpka_del() {
+	echo "tcpka_del ..."
 	while true
 	do
 		IPID=`dhd_priv wl tcpka_conn_sess_info 1 | grep ipid`
@@ -110,23 +155,31 @@ function tcpka_del() {
 		else
 			break
 		fi
+
+		let tcpka_cnt++
+		if [ $tcpka_cnt -gt 3 ]; then
+			echo "tcpka_conn_del failed"
+			exit
+		fi
+
 	done
 }
 
 function wlan_up() {
+	echo "wlan up ..."
 	while true
 	do
 		ifconfig wlan0 up
 		if [ $? -ne 0 ]; then
 			echo "wlan0 up failed"
-			sleep 0.2
+			sleep 0.1
 		else
 			echo "wlan0 up succeed"
 			break
 		fi
 
 		let up_cnt++
-		if [ $up_cnt -gt 10 ]; then
+		if [ $up_cnt -gt 3 ]; then
 			echo "wlan0 up failed!!!"
 			echo 0 > /sys/class/leds/blue/brightness
 			exit
@@ -135,13 +188,30 @@ function wlan_up() {
 }
 
 function connect_wifi() {
-
+echo "connect wifi ..."
 SSID=`dhd_priv isam_status | awk -F, '{print $5}' | sed -n '4p' | cut -d '"' -f 2`
 
 if [ "$SSID" ==  "" ] || [ "$FORCE_CONFIG_WIFI" == "true" ];then
 
-	if [ "$SSID" ==  "" ] && [ "$FORCE_CONFIG_WIFI" != "true" ];then
-		echo "SSID is empty, and FORCE_CONFIG_WIFI not true"
+	if [ "$FORCE_CONFIG_WIFI" != "true" ];then
+
+		VVAILD=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $1}'`
+		echo "vvaild: $VVAILD"
+		VSSID=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $2}'`
+		echo "vssid: $VSSID"
+		VPSK=`vendor_storage -r VENDOR_CUSTOM_ID_1E -t string |  sed -n '4p' | awk '{print $3}' | awk -F, '{print $3}'`
+		echo "vpsk: $VPSK"
+
+		if [ "$VVAILD" == "1" ];then
+			WIFISSID=$VSSID
+			WIFIPWD=$VPSK
+		fi
+
+		echo "vendor storage: vaild: $VVAILD ssid: $WIFISSID, psk: $WIFIPWD"
+	fi
+
+	if [ "$SSID" ==  "" ] && [ "$FORCE_CONFIG_WIFI" != "true" ] && [ "$VVAILD" != "1" ];then
+		echo "SSID is empty, and FORCE_CONFIG_WIFI not true and vendor is empty, exit"
 		echo 0 > /sys/class/leds/blue/brightness
 		exit
 	fi
@@ -173,14 +243,14 @@ if [ "$SSID" ==  "" ] || [ "$FORCE_CONFIG_WIFI" == "true" ];then
 		wpa_supplicant -B -i wlan0 -c /tmp/wpa_supplicant.conf
 		WPAPID=`busybox ps | grep wpa_supplicant | grep -v grep | awk '{print $1}'`
 		if [ "$WPAPID" == "" ];then
-			sleep 0.5
+			sleep 0.2
 		else
 			echo "wpa_supplicant running"
 			break
 		fi
 
 		let wpa_cnt++
-		if [ $wpa_cnt -gt 10 ]; then
+		if [ $wpa_cnt -gt 3 ]; then
 			echo "wpa_supplicant up failed!!!"
 			echo 0 > /sys/class/leds/blue/brightness
 			exit
@@ -196,13 +266,14 @@ if [ "$SSID" ==  "" ] || [ "$FORCE_CONFIG_WIFI" == "true" ];then
 		elif [ "$NEW_SSID" == "$SSID" ];then
 			echo "waiting $SSID disconnect ..."
 			sleep 1
-		else [ "$NEW_SSID" == "WIFISSID" ]
+		else [ "$NEW_SSID" == "$WIFISSID" ]
 			echo "$NEW_SSID connected to getdhcp"
+			NET_NEW_CONFIG=1
 			break
 		fi
 
 		let connect_cnt++
-		if [ $connect_cnt -gt 60 ]; then
+		if [ $connect_cnt -gt 35 ]; then
 			echo "connect $NEW_SSID failed!!!"
 			echo 0 > /sys/class/leds/blue/brightness
 			exit
@@ -215,4 +286,4 @@ check_wlan0
 wlan_up
 tcpka_del
 connect_wifi
-getdhcp
+udhcpcd
